requirements.txt
pynacl
pysha3
pyqt6